Refactor: Correctly name Facade classes from wrappers, add config
