+++ src/AudioChord/MusicServiceConfiguration.cs
@@ -1,8 +1,12 @@
 ﻿using AudioChord.Caching;
+using AudioChord.Extractors;
 using System;
 
 namespace AudioChord
 {
+    /// <summary>
+    /// Configuration object for the music service
+    /// </summary>
     public class MusicServiceConfiguration
     {
         public Func<ISongCache> SongCacheFactory { get; set; }
@@ -13,5 +17,7 @@ namespace AudioChord
 
         // ReSharper disable once StringLiteralTypo
         public string Database { get; internal set; } = "sharedmusic";
+        
+        public ExtractorConfiguration ExtractorConfiguration { get; set; } = new ExtractorConfiguration();
     }
 }
